feat(core): added the GameMaster
the game master manages the state and execution of the game for the
purpose of abstracting the inner workings of the game from the ui. the ui
layer will simply pass inputs to the game master, who will run through
the game logic to generate a result, and hand the result to the ui to
generate output. this way, the ui layer never needs to understand the
underlying business process of running the game, only the flow of a
game.

currently, the game master can run the betting and standoff phase. also
made the following changes to support:

- include constructor with message for NotEnoughCashException
- add compareTo method to Card